+++ lib/src/extensions/iterable.dart
@@ -50,6 +50,25 @@ Would you like any specific implementation details or examples for any of these 
 typedef IndexedPredicate<T> = bool Function(int index, T);
 typedef Predicate<T> = bool Function(T);
 
+extension DHUNullableSetExtensions<E> on Set<E>? {
+  /// Converts the set to a different type [R], similar to the original [cast] method,
+  /// but with additional flexibility and error handling.
+  ///
+  /// Unlike the standard [Set.cast], which can throw a [CastError] if an element cannot
+  /// be cast to the specified type [R], [castTo] leverages a custom conversion logic
+  /// which tries various strategies to convert elements.
+  ///
+  /// If a direct cast is not feasible, it attempts element-wise conversion using `toType`.
+  /// This ensures that the conversion happens smoothly without risking a crash.
+  ///
+  /// Example usage:
+  /// ```dart
+  /// Set<dynamic> set = {1, 2, '3'};
+  /// List<int> intList = set.castTo<int>(); // Tries to convert all elements to int.
+  /// ```
+  Set<R> castTo<R>() => ConvertObject.toSet<R>(this);
+}
+
 extension DHUNullableListExtensions<E> on List<E>? {
   /// same behavior as [removeAt] but it is null safe which means
   /// it do nothing when [List] return [isEmptyOrNull] to true.
@@ -80,6 +99,23 @@ extension DHUNullableListExtensions<E> on List<E>? {
   /// it do nothing when [List] return [isEmptyOrNull] to true.
   void tryRemoveWhere(int element) =>
       isEmptyOrNull ? null : this!.removeWhere((element) => false);
+
+  /// Converts the list to a different type [R], similar to the original [cast] method,
+  /// but with additional flexibility and error handling.
+  ///
+  /// Unlike the standard [List.cast], which can throw a [CastError] if an element cannot
+  /// be cast to the specified type [R], [castTo] leverages a custom conversion logic
+  /// which tries various strategies to convert elements.
+  ///
+  /// If a direct cast is not feasible, it attempts element-wise conversion using `toType`.
+  /// This ensures that the conversion happens smoothly without risking a crash.
+  ///
+  /// Example usage:
+  /// ```dart
+  /// List<dynamic> list = [1, 2, '3'];
+  /// List<int> intList = list.castTo<int>(); // Tries to convert all elements to int.
+  /// ```
+  List<R> castTo<R>() => ConvertObject.toList<R>(this);
 }
 
 extension DHUCollectionsExtensionsNS<E> on Iterable<E>? {
@@ -361,6 +397,32 @@ extension DHUCollectionsExtensionsNS<E> on Iterable<E>? {
 }
 
 extension DHUCollectionsExtensions<E> on Iterable<E> {
+  /// Converts the iterable to a [List] of type [R] using the [castTo] method.
+  ///
+  /// This method first converts the iterable to a list, then casts it to the desired
+  /// type [R] using custom conversion logic provided by [castTo]. This ensures flexibility
+  /// and handles potential casting errors gracefully.
+  ///
+  /// Example usage:
+  /// ```dart
+  /// Iterable<dynamic> iterable = [1, '2', 3.0];
+  /// List<int> intList = iterable.toListCasted<int>(); // Tries to convert all elements to int.
+  /// ```
+  List<R> toListCasted<R>() => this.toList().castTo<R>();
+
+  /// Converts the iterable to a [Set] of type [R] using the [castTo] method.
+  ///
+  /// This method first converts the iterable to a set, then casts it to the desired
+  /// type [R] using custom conversion logic provided by [castTo]. This ensures flexibility
+  /// and handles potential casting errors gracefully.
+  ///
+  /// Example usage:
+  /// ```dart
+  /// Iterable<dynamic> iterable = [1, '2', 3.0];
+  /// Set<int> intSet = iterable.toSetCasted<int>(); // Tries to convert all elements to int.
+  /// ```
+  Set<R> toSetCasted<R>() => this.toSet().castTo<R>();
+
   /// Returns this Iterable if it's not `null` and the empty list otherwise.
   Iterable<E> orEmpty() => this;